fix. CaptureActivity
캡쳐 관련 코드를 별도의 유틸리티로 분리
