+++ presentation/src/main/java/com/ddd4/dropit/presentation/ui/capture/CaptureActivity.kt
@@ -1,18 +1,7 @@
 package com.ddd4.dropit.presentation.ui.capture
 
-import android.annotation.SuppressLint
 import android.app.Activity
-import android.content.Context
 import android.graphics.*
-import android.hardware.camera2.*
-import android.media.Image
-import android.media.ImageReader
-import android.net.Uri
-import android.os.Bundle
-import android.provider.MediaStore
-import android.util.Size
-import android.util.SparseIntArray
-import android.view.Surface
 import android.view.TextureView
 import androidx.activity.viewModels
 import androidx.lifecycle.Observer
@@ -20,43 +9,24 @@ import com.ddd4.dropit.presentation.R
 import com.ddd4.dropit.presentation.base.ui.BaseActivity
 import com.ddd4.dropit.presentation.databinding.ActivityCaptureBinding
 import com.ddd4.dropit.presentation.util.Constants
-import com.ddd4.dropit.presentation.util.toast
+import com.ddd4.dropit.presentation.util.capture.TextureCaptureUtil
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.android.synthetic.main.activity_add.toolbar
 import kotlinx.android.synthetic.main.activity_capture.*
-import timber.log.Timber
-import java.io.*
-import java.text.SimpleDateFormat
-import java.util.*
-import kotlin.collections.ArrayList
-
+import javax.inject.Inject
 
 @AndroidEntryPoint
 class CaptureActivity : BaseActivity<ActivityCaptureBinding>(R.layout.activity_capture) {
 
     private val captureViewModel: CaptureViewModel by viewModels()
 
+    @Inject
+    lateinit var textureCaptureUtil: TextureCaptureUtil
+
     override fun setBind() {
         binding.apply {
+            view = this@CaptureActivity
             captureVM = captureViewModel
-        }
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        setupActionBar()
-
-        texture.surfaceTextureListener = textureListener
-    }
-
-    private fun setupActionBar() {
-        setSupportActionBar(toolbar)
-        supportActionBar.apply {
-            this?.setDisplayShowHomeEnabled(true)
-            this?.setDisplayHomeAsUpEnabled(true)
-            this?.setHomeAsUpIndicator(R.drawable.ic_back_gray_32dp)
-            this?.setDisplayShowTitleEnabled(false)
         }
     }
 
@@ -66,239 +36,40 @@ class CaptureActivity : BaseActivity<ActivityCaptureBinding>(R.layout.activity_c
     }
 
     override fun setObserve() {
-        captureViewModel.capturedImage.observe(this, Observer { imagePath ->
-            val intent = intent.putExtra(Constants.EXTRA_NAME_IMAGE_PATH, imagePath)
+        captureViewModel.captureComplete.observe(this, Observer { imageUri ->
+            val intent = intent.putExtra(Constants.EXTRA_NAME_IMAGE_PATH, imageUri)
             setResult(Activity.RESULT_OK, intent)
             finish()
         })
         captureViewModel.captureClick.observe(this, Observer {
-            takePicture()
+            textureCaptureUtil.takePicture(this, setImageListener)
         })
     }
 
-    private val textureListener = object: TextureView.SurfaceTextureListener {
+    override fun setInit() {
+        setSupportActionBar(toolbar)
+        supportActionBar.apply {
+            this?.setDisplayShowHomeEnabled(true)
+            this?.setDisplayHomeAsUpEnabled(true)
+            this?.setHomeAsUpIndicator(R.drawable.ic_back_gray_32dp)
+            this?.setDisplayShowTitleEnabled(false)
+        }
+    }
+
+    val textureListener = object: TextureView.SurfaceTextureListener {
         override fun onSurfaceTextureSizeChanged(surface: SurfaceTexture?, width: Int, height: Int) {}
         override fun onSurfaceTextureUpdated(surface: SurfaceTexture?) {}
         override fun onSurfaceTextureDestroyed(surface: SurfaceTexture?): Boolean {
             return false
         }
         override fun onSurfaceTextureAvailable(surface: SurfaceTexture?, width: Int, height: Int) {
-            openCamera()
+            textureCaptureUtil.openCamera(texture)
         }
     }
 
-    private lateinit var cameraId: String
-    private lateinit var imageDimension: Size
-    private var cameraDevice: CameraDevice? = null
-
-    @SuppressLint("MissingPermission")
-    private fun openCamera() {
-        // 카메라의 정보를 가져와서 cameraId 와 imageDimension 에 값을 할당하고,
-        // 카메라를 열어야 하기 때문에 CameraManager 객체를 가져온다
-        val manager = getSystemService(Context.CAMERA_SERVICE) as CameraManager
-
-        try {
-            // CameraManager 에서 cameraIdList 의 값을 가져온다
-            // FaceCamera 값이 true 이면 전면, 아니면 후면 카메라
-            cameraId = manager.cameraIdList[0]
-
-            val characteristics = manager.getCameraCharacteristics(cameraId)
-            val map = characteristics.get(CameraCharacteristics.SCALER_STREAM_CONFIGURATION_MAP)
-
-            // SurfaceTexture 에 사용할 Size 값을 map 에서 가져와 imageDimension 에 할당해준다
-            imageDimension = map!!.getOutputSizes<SurfaceTexture>(SurfaceTexture::class.java)[0]
-
-            // CameraManager.openCamera() 메서드를 이용해 인자로 넘겨준 cameraId 의 카메라를 실행한다
-            // 이때, stateCallback 은 카메라를 실행할때 호출되는 콜백메서드이며, cameraDevice 에 값을 할달해주고, 카메라 미리보기를 생성한다
-            manager.openCamera(cameraId, stateCallback, null)
-
-        } catch (e: CameraAccessException) {
-            e.printStackTrace()
+    private val setImageListener = object: TextureCaptureUtil.OnTextureImageListener {
+        override fun onImageCaptured(imageUri: String) {
+            captureViewModel.setCapturedImage(imageUri)
         }
-    }
-
-    private val stateCallback = object : CameraDevice.StateCallback() {
-        override fun onOpened(camera: CameraDevice) {
-            Timber.d("stateCallback : onOpened")
-
-            // MainActivity 의 cameraDevice 에 값을 할당해주고, 카메라 미리보기를 시작한다
-            // 나중에 cameraDevice 리소스를 해지할때 해당 cameraDevice 객체의 참조가 필요하므로,
-            // 인자로 들어온 camera 값을 전역변수 cameraDevice 에 넣어 준다
-            cameraDevice = camera
-
-            // createCameraPreview() 메서드로 카메라 미리보기를 생성해준다
-            createCameraPreviewSession()
-        }
-        override fun onDisconnected(camera: CameraDevice) {
-            Timber.d("stateCallback : onDisconnected")
-
-            // 연결이 해제되면 cameraDevice 를 닫아준다
-            cameraDevice!!.close()
-        }
-        override fun onError(camera: CameraDevice, error: Int) {
-            Timber.d("stateCallback : onError")
-
-            // 에러가 뜨면, cameraDevice 를 닫고, 전역변수 cameraDevice 에 null 값을 할당해 준다
-            cameraDevice!!.close()
-            cameraDevice = null
-        }
-    }
-
-    private fun createCameraPreviewSession() {
-        try {
-            // 캡쳐세션을 만들기 전에 프리뷰를 위한 Surface 를 준비한다
-            // 레이아웃에 선언된 textureView 로부터 surfaceTexture 를 얻을 수 있다
-            val texture = texture.surfaceTexture
-
-            // 미리보기를 위한 Surface 기본 버퍼의 크기는 카메라 미리보기크기로 구성
-            texture.setDefaultBufferSize(imageDimension.width, imageDimension.height)
-
-            // 미리보기를 시작하기 위해 필요한 출력표면인 surface
-            val surface = Surface(texture)
-
-            // 미리보기 화면을 요청하는 RequestBuilder 를 만들어준다.
-            // 이 요청은 위에서 만든 surface 를 타겟으로 한다
-            val captureRequestBuilder = cameraDevice!!.createCaptureRequest(CameraDevice.TEMPLATE_PREVIEW)
-            captureRequestBuilder.addTarget(surface)
-
-            // 위에서 만든 surface 에 미리보기를 보여주기 위해 createCaptureSession() 메서드를 시작한다
-            // createCaptureSession 의 콜백메서드를 통해 onConfigured 상태가 확인되면
-            // CameraCaptureSession 을 통해 미리보기를 보여주기 시작한다
-            cameraDevice!!.createCaptureSession(listOf(surface), object : CameraCaptureSession.StateCallback() {
-                override fun onConfigureFailed(session: CameraCaptureSession) {}
-                override fun onConfigured(session: CameraCaptureSession) {
-                    if(cameraDevice == null) {
-                        // 카메라가 이미 닫혀있는경우, 열려있지 않은 경우
-                        return
-                    }
-                    // session 이 준비가 완료되면, 미리보기를 화면에 뿌려주기 시작한다
-                    //captureRequestBuilder.set(CaptureRequest.CONTROL_MODE, CameraMetadata.CONTROL_MODE_AUTO)
-                    val cropRect: Rect = Rect(0, 0, 1755, 3120)
-                    captureRequestBuilder.set(CaptureRequest.SCALER_CROP_REGION, cropRect)
-
-                    try {
-                        session.setRepeatingRequest(captureRequestBuilder.build(), null, null)
-                    } catch (e: CameraAccessException) {
-                        e.printStackTrace()
-                    }
-                }
-            }, null)
-        } catch (e: CameraAccessException) {
-            e.printStackTrace()
-        }
-    }
-
-    private var fileCount: Int = 0
-
-    private fun takePicture() {
-        try {
-            val manager = getSystemService(Context.CAMERA_SERVICE) as CameraManager
-            val characteristics = manager.getCameraCharacteristics(cameraDevice!!.id)
-            val jpegSizes = characteristics.get(CameraCharacteristics.SCALER_STREAM_CONFIGURATION_MAP)!!.getOutputSizes(ImageFormat.JPEG)
-
-            val width = jpegSizes[0].width
-            val height = jpegSizes[0].height
-
-            val imageReader = ImageReader.newInstance(width, height, ImageFormat.JPEG, 1)
-
-            val outputSurface = ArrayList<Surface>(2)
-            outputSurface.add(imageReader.surface)
-            outputSurface.add(Surface(texture!!.surfaceTexture))
-
-            val captureBuilder = cameraDevice!!.createCaptureRequest(CameraDevice.TEMPLATE_STILL_CAPTURE)
-            captureBuilder.addTarget(imageReader.surface)
-
-            captureBuilder.set(CaptureRequest.CONTROL_MODE, CameraMetadata.CONTROL_MODE_AUTO)
-
-            val sparseIntArray = SparseIntArray()
-            // 사진의 rotation 을 설정해준다
-            val rotation = windowManager.defaultDisplay.rotation
-            captureBuilder.set(CaptureRequest.JPEG_ORIENTATION, sparseIntArray.get(rotation))
-
-            val timeStamp = SimpleDateFormat("yyyyMMddHHmmss", Locale.KOREA).format(Date())
-            val imageName = "${timeStamp}_"
-            val storageDir = this.filesDir
-            val imageFile = File.createTempFile(imageName, ".jpg", storageDir)
-
-            val readerListener = ImageReader.OnImageAvailableListener {
-                var image : Image? = null
-
-                try {
-                    image = imageReader.acquireLatestImage()
-
-                    val buffer = image!!.planes[0].buffer
-                    val bytes = ByteArray(buffer.capacity())
-                    buffer.get(bytes)
-
-                    var output: OutputStream? = null
-                    try {
-                        output = FileOutputStream(imageFile)
-                        output.write(bytes)
-                    } finally {
-                        output?.close()
-
-                        val uri = Uri.fromFile(imageFile)
-                        Timber.d("Image Uri Path $uri")
-
-                        // 프리뷰 이미지에 set 해줄 비트맵을 만들어준다
-                        val bitmap: Bitmap = BitmapFactory.decodeFile(imageFile.path)
-
-                        // 비트맵 사진이 90도 돌아가있는 문제를 해결하기 위해 rotate 해준다
-                        val rotateMatrix = Matrix()
-                        rotateMatrix.postRotate(90F)
-                        val rotatedBitmap = Bitmap.createBitmap(bitmap, 0,0, bitmap.width, bitmap.height, rotateMatrix, false)
-
-                        // 90도 돌아간 비트맵을 이미지뷰에 set 해준다
-                        //Timber.d("이미지 비트맵: $rotatedBitmap")
-                        captureViewModel.capturedImage.value = getImageUri(rotatedBitmap).toString()
-
-                        fileCount++
-                    }
-                } catch (e: FileNotFoundException) {
-                    e.printStackTrace()
-                } catch (e: IOException) {
-                    e.printStackTrace()
-                } finally {
-                    image?.close()
-                }
-            }
-
-            // imageReader 객체에 위에서 만든 readerListener 를 달아서, 이미지가 사용가능하면 사진을 저장한다
-            imageReader.setOnImageAvailableListener(readerListener, null)
-
-            val captureListener = object : CameraCaptureSession.CaptureCallback() {
-                override fun onCaptureCompleted(session: CameraCaptureSession, request: CaptureRequest, result: TotalCaptureResult) {
-                    super.onCaptureCompleted(session, request, result)
-                    toast("이미지를 가져왔습니다")
-                    createCameraPreviewSession()
-                }
-            }
-
-            // outputSurface 에 위에서 만든 captureListener 를 달아,
-            // 캡쳐(사진 찍기) 해주고 나서 카메라 미리보기 세션을 재시작한다
-            cameraDevice!!.createCaptureSession(outputSurface, object : CameraCaptureSession.StateCallback() {
-                override fun onConfigureFailed(session: CameraCaptureSession) {}
-
-                override fun onConfigured(session: CameraCaptureSession) {
-                    try {
-                        session.capture(captureBuilder.build(), captureListener, null)
-                    } catch (e: CameraAccessException) {
-                        e.printStackTrace()
-                    }
-                }
-
-            }, null)
-        } catch (e: CameraAccessException) {
-            e.printStackTrace()
-        }
-    }
-
-    private fun getImageUri(bitmap: Bitmap): Uri? {
-        val bytes = ByteArrayOutputStream()
-        bitmap.compress(Bitmap.CompressFormat.JPEG, 100, bytes)
-        val titleFormat = SimpleDateFormat("yyyyMMddHHmmss", Locale.KOREA)
-        val title = titleFormat.format(Date())
-        val path = MediaStore.Images.Media.insertImage(contentResolver, bitmap, title, null)
-        return Uri.parse(path)
     }
 }
